refactor (schemaExporter): Now is typed as abstract
